+++ src/components/settings/SettingsAutoSaveToField.jsx
@@ -1,12 +1,12 @@
-import SettingsSelectField from './SettingsSelectField';
+import SettingsFieldSelect from './SettingsFieldSelect';
 
 export default function SettingsAutoSaveToField() {
   return (
-    <SettingsSelectField title="Auto-save to:" id="save-to">
+    <SettingsFieldSelect title="Auto-save to:" id="save-to">
       <option value="local">
         {'<'}This local device{'>'}
       </option>
       <option value="gdrive">Google Drive</option>
-    </SettingsSelectField>
+    </SettingsFieldSelect>
   );
 }