+++ src/components/settings/SettingsEnableGoogleDriveSyncToggle.jsx
@@ -1,16 +1,14 @@
-import { MenuItem } from '@ariakit/react';
 import { useCallback, useEffect, useState } from 'react';
 
 import AddToDriveIcon from '@material-symbols/svg-400/rounded/add_to_drive.svg';
-import ToggleOff from '@material-symbols/svg-400/rounded/toggle_off.svg';
-import ToggleOn from '@material-symbols/svg-400/rounded/toggle_on-fill.svg';
 
 import { useInterval } from '@/libs/UseInterval';
 import { useGAPITokenHandler } from '@/libs/googleapi';
 import { useSettingsStore } from '@/stores/settings';
-import MenuStyle from '@/styles/Menu.module.css';
 
-export default function EnableThumbnailWhileRecording() {
+import SettingsFieldToggle from './SettingsFieldToggle';
+
+export default function SettingsEnableGoogleDriveSyncToggle() {
   const [loggedIn, setLoggedIn] = useState(false);
   const enabled = useSettingsStore((ctx) => ctx.user.enableDriveSync);
   const setEnabled = useSettingsStore((ctx) => ctx.setEnableDriveSync);
@@ -25,18 +23,14 @@ export default function EnableThumbnailWhileRecording() {
   useEffect(onInterval, [onInterval]);
 
   return (
-    <MenuItem
-      className={MenuStyle.menuItem + ' ' + 'flex flex-row fill-current'}
-      hideOnClick={false}
+    <SettingsFieldToggle
+      className={'flex flex-row fill-current'}
+      value={enabled}
       onClick={() => setEnabled(!enabled)}
       disabled={!loggedIn}>
-      {enabled ? (
-        <ToggleOn className="h-full" />
-      ) : (
-        <ToggleOff className="h-full" />
-      )}
-      Enable
-      <AddToDriveIcon className="w-6 h-6 fill-current" /> Drive Sync
-    </MenuItem>
+      <span>Enable</span>
+      <AddToDriveIcon className="inline-block w-6 h-6 fill-current mx-1" />
+      <span>Drive sync</span>
+    </SettingsFieldToggle>
   );
 }