+++ src/components/settings/SettingsVideoResolutionField.jsx
@@ -1,10 +1,10 @@
-import SettingsSelectField from './SettingsSelectField';
+import SettingsFieldSelect from './SettingsFieldSelect';
 
 export default function SettingsVideoResolutionField() {
   return (
-    <SettingsSelectField title="Video Resolution:" id="video-resolution">
+    <SettingsFieldSelect title="Video Resolution:" id="video-resolution">
       <option value="4K">4K</option>
       <option value="HD">HD</option>
-    </SettingsSelectField>
+    </SettingsFieldSelect>
   );
 }