+++ src/routes/SettingsPage.jsx
@@ -1,9 +1,19 @@
 import NavBarLayout from '@/components/NavBarLayout';
 import SettingsAspectRatioField from '@/components/settings/SettingsAspectRatioField';
 import SettingsAutoSaveToField from '@/components/settings/SettingsAutoSaveToField';
+import SettingsEnableGoogleDriveSyncToggle from '@/components/settings/SettingsEnableGoogleDriveSyncToggle';
+import SettingsEnableRecorderLiveAudioToggle from '@/components/settings/SettingsEnableRecorderLiveAudioToggle';
+import SettingsEnableRecorderReferenceToggle from '@/components/settings/SettingsEnableRecorderReferenceToggle';
+import SettingsFieldGroup from '@/components/settings/SettingsFieldGroup';
+import SettingsNerdInfoButton from '@/components/settings/SettingsNerdInfoButton';
+import SettingsPreferNativeRecorderToggle from '@/components/settings/SettingsPreferNativeRecorderToggle';
+import SettingsPreferPersistedMediaStreamToggle from '@/components/settings/SettingsPreferPersistedMediaStreamToggle';
+import SettingsProjectDeleteButton from '@/components/settings/SettingsProjectDeleteButton';
 import SettingsProjectIdField from '@/components/settings/SettingsProjectIdField';
 import SettingsProjectNameField from '@/components/settings/SettingsProjectNameField';
 import SettingsReturnHomeField from '@/components/settings/SettingsReturnHomeField';
+import SettingsShareFilesButton from '@/components/settings/SettingsShareFilesButton';
+import SettingsVideoCacheClearButton from '@/components/settings/SettingsVideoCacheClearButton';
 import SettingsVideoResolutionField from '@/components/settings/SettingsVideoResolutionField';
 
 export default function SettingsPage() {
@@ -27,6 +37,23 @@ export default function SettingsPage() {
           <SettingsAutoSaveToField />
           <br />
           <SettingsReturnHomeField />
+          <br />
+          <SettingsFieldGroup title="Data Settings">
+            <SettingsEnableGoogleDriveSyncToggle />
+            <SettingsShareFilesButton />
+          </SettingsFieldGroup>
+          <br />
+          <SettingsFieldGroup title="Recorder Settings">
+            <SettingsPreferNativeRecorderToggle />
+            <SettingsPreferPersistedMediaStreamToggle />
+            <SettingsEnableRecorderReferenceToggle />
+            <SettingsEnableRecorderLiveAudioToggle />
+          </SettingsFieldGroup>
+          <br />
+          <SettingsNerdInfoButton />
+          <br />
+          <SettingsVideoCacheClearButton />
+          <SettingsProjectDeleteButton />
           <br />
         </fieldset>
       </NavBarLayout>