+++ src/stores/document/dispatch/DispatchDocumentSettings.js
@@ -27,12 +27,7 @@ export function createDispatchDocumentSettings(set, get) {
  * @param {string} projectId
  */
 function setDocumentSettingsProjectId(store, documentId, projectId) {
-  let settings = getDocumentSettingsById(store, documentId);
-  if (!settings) {
-    // @ts-ignore
-    settings = {};
-    getDocumentById(store, documentId).settings = settings;
-  }
+  let settings = resolveDocumentSettingsById(store, documentId);
   if (projectId && projectId.trim().length > 0) {
     settings.projectId = formatProjectId(projectId);
   } else {
@@ -46,12 +41,7 @@ function setDocumentSettingsProjectId(store, documentId, projectId) {
  * @param {'16:9'|'4:3'} aspectRatio
  */
 function setDocumentSettingsAspectRatio(store, documentId, aspectRatio) {
-  let settings = getDocumentSettingsById(store, documentId);
-  if (!settings) {
-    // @ts-ignore
-    settings = {};
-    getDocumentById(store, documentId).settings = settings;
-  }
+  let settings = resolveDocumentSettingsById(store, documentId);
   settings.aspectRatio = aspectRatio;
 }
 
@@ -65,11 +55,21 @@ function setDocumentSettingsVideoResolution(
   documentId,
   videoResolution,
 ) {
+  let settings = resolveDocumentSettingsById(store, documentId);
+  settings.videoResolution = videoResolution;
+}
+
+/**
+ * @param {import('../DocumentStore').Store} store
+ * @param {import('../DocumentStore').DocumentId} documentId
+ */
+function resolveDocumentSettingsById(store, documentId) {
+  /** @type {Partial<import('../DocumentStore').Document['settings']>} */
   let settings = getDocumentSettingsById(store, documentId);
   if (!settings) {
-    // @ts-ignore
     settings = {};
+    // @ts-expect-error settings should always be able to be partially constructed.
     getDocumentById(store, documentId).settings = settings;
   }
-  settings.videoResolution = videoResolution;
+  return settings;
 }